requirements.txt
asgiref==3.7.2
boto3
certifi==2024.2.2
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django>=3.2,<4.0
django-model-utils==4.4.0
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
geopy
idna==3.6
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
openapi-codec==1.3.2
packaging==24.0
psycopg2
pillow>=8.0
PyJWT==2.8.0
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
simplejson==3.19.2
sqlparse==0.4.4
Unidecode==1.3.8
uritemplate==4.1.1
urllib3==2.2.1
django-cors-headers
pandas
pdfkit
python-docx
beautifulsoup4


